requirements.txt
gunicorn
certifi==2017.11.5
click==6.7
Flask==0.12.2
Flask-Cors
Flask-Login==0.4.0
Flask-SocketIO==2.9.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
nltk==3.2.5
numpy==1.13.3
pandas==0.21.0
python-dateutil==2.6.1
python-engineio==2.0.1
python-socketio==1.8.3
pytz==2017.3
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
Werkzeug==0.12.2
paho-mqtt
requests
textblob
psycopg2
scikit-image
dlib
psycopg2-binary
